requirement.txt
flask==2.0.3
flask-mysqldb==1.0.1
python-dotenv==0.19.2